reorganization & fix output file writing
